% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facet_layout.R
\name{facet_layout}
\alias{facet_layout}
\title{Forced layout in \code{facet_wrap()}}
\usage{
facet_layout(plot = NULL, facets = NULL, nrow = 2, ncol = 2,
  scales = "fixed")
}
\arguments{
\item{plot}{a ggplot object}

\item{facets}{variables to facet by}

\item{nrow}{number of rows}

\item{ncol}{number of columns}

\item{scales}{should scales be fixed ("fixed", the default), free ("free"),
or free in one dimension ("free_x", "free_y")}
}
\description{
Allows \code{ggplot2} to use bigger layout than usually possible in
\code{facet_wrap()} given the number of panels.
}
\examples{
\dontrun{
p <- ggplot(diamonds, aes(x = price, y = carat, color = cut)) +
geom_point(alpha = 0.5) +
labs(x = 'Price', y = 'Carat', title = 'Diamonds')

facet_layout(plot = p, facets = 'color', ncol = 4, nrow = 3)
}
}
\seealso{
\code{\link{facet_wrap}}, \code{\link{facet_multiple}}
}

