% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_readable.R
\name{theme_readable}
\alias{theme_readable}
\title{ggplot2 theme}
\usage{
theme_readable(base_size = 12, legend_position = "right")
}
\arguments{
\item{base_size}{base font size}

\item{legend_position}{position of the legend ("none", "left", "right", "bottom", 
"top", or two-element numeric vector)}
}
\description{
Theme intended to make ggplot2 more readable when used 
in presentation or papers. Background and major grid lines were dimed 
and minor grid lines removed to focus the attention on the data.
}
\examples{
\dontrun{
ggplot(diamonds, aes(x = price, y = carat, color = cut)) + 
 geom_point(alpha = 0.5) + 
 theme_readable(legend_position = 'top')
}
}
\seealso{
\code{\link{ggtheme}}, \code{\link{theme}}
}

